AMP-19384. Code clean up and improvements
git-svn-id: file:///svn/branches/AMP_2_11_RELEASE@26756 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/aim/util/AidEffectivenessIndicatorUtil.java
@@ -1,9 +1,11 @@
 package org.digijava.module.aim.util;
 
+import org.apache.log4j.Logger;
 import org.digijava.kernel.persistence.PersistenceManager;
 import org.digijava.module.aim.dbentity.AmpActivityVersion;
 import org.digijava.module.aim.dbentity.AmpAidEffectivenessIndicator;
 import org.digijava.module.aim.dbentity.AmpAidEffectivenessIndicatorOption;
+import org.hibernate.ObjectNotFoundException;
 import org.hibernate.Query;
 import org.hibernate.Session;
 
@@ -13,6 +15,8 @@ import java.util.*;
  * List of utility methods to operate with effectiveness indicators
  */
 public class AidEffectivenessIndicatorUtil {
+
+    private static Logger logger = Logger.getLogger(AidEffectivenessIndicatorUtil.class);
 
     /**
      * Returns list of indicators by the criteria
@@ -55,14 +59,38 @@ public class AidEffectivenessIndicatorUtil {
         return  (List<AmpAidEffectivenessIndicator>)query.list();
     }
 
+    /**
+     * Loads indicator by primary key from the database
+     * @param indicatorId
+     * @ throws org.hibernate.ObjectNotFoundException if the object was not found in the database,
+     * usually because the id is incorrect
+     * @return AmpAidEffectivenessIndicator object, null if object was not found
+     */
     public static AmpAidEffectivenessIndicator loadIndicatorById(Long indicatorId) {
         Session session = PersistenceManager.getSession();
-        return (AmpAidEffectivenessIndicator) session.load(AmpAidEffectivenessIndicator.class, indicatorId);
+        try {
+            return (AmpAidEffectivenessIndicator) session.load(AmpAidEffectivenessIndicator.class, indicatorId);
+        } catch (ObjectNotFoundException nfe) {
+            logger.error(nfe.getStackTrace());
+            return null;
+        }
     }
 
+    /**
+     * Loads indicator option by primary key from the database
+     * @param optionId
+     * @ throws org.hibernate.ObjectNotFoundException if the object was not found in the database,
+     * usually because the id is incorrect
+     * @return AmpAidEffectivenessIndicatorOption object, null if object was not found
+     */
     public static AmpAidEffectivenessIndicatorOption loadOptionById(Long optionId) {
         Session session = PersistenceManager.getSession();
-        return (AmpAidEffectivenessIndicatorOption) session.load(AmpAidEffectivenessIndicatorOption.class, optionId);
+        try {
+            return (AmpAidEffectivenessIndicatorOption) session.load(AmpAidEffectivenessIndicatorOption.class, optionId);
+        } catch (ObjectNotFoundException nfe) {
+            logger.error(nfe.getStackTrace());
+            return null;
+        }
     }
 
     public static void saveIndicator(AmpAidEffectivenessIndicator indicator) {
@@ -74,29 +102,28 @@ public class AidEffectivenessIndicatorUtil {
      * Do not delete via Id, but load first
      * Because we want to delete related options as well
      * @param indicatorId the indicator Id to delete
+     * @return returns null if indicator with the id provided was not found in the database
      */
-    public static void deleteIndicator(long indicatorId) {
+    public static AmpAidEffectivenessIndicator deleteIndicator(long indicatorId) {
         Session session = PersistenceManager.getSession();
         AmpAidEffectivenessIndicator indicator = null;
-        try {
-            indicator = loadIndicatorById(indicatorId);
-        }  catch (org.hibernate.ObjectNotFoundException nfe) {
-            /*
-             * Throwing Runtime here to prevent handling Hibernate exceptions in the action class
-             * I do not want to create business level exception for this
-             */
-            throw new RuntimeException("Indicator with id " + indicatorId + " not found");
-        }
 
-        // these two methods should be executed withing the transaction context
-        session.delete(indicator);
-        cleanUpModulesVisibility(indicator);
+        indicator = loadIndicatorById(indicatorId);
+        if (indicator != null) {
+            // these two methods should be executed withing the transaction context
+            session.delete(indicator);
+            cleanUpModulesVisibility(indicator);
+            return indicator;
+        } else {
+            return null;
+        }
     }
 
     /**
      * Delete option and the relationship with indicator
      * @param optionId the option id to be deleted
      * @return returns the indicator for data be refreshed after deletion
+     * returns null if indicator with the id provided was not found in the database
      */
     public static AmpAidEffectivenessIndicator deleteOption(long optionId) {
         Session session = PersistenceManager.getSession();
@@ -105,11 +132,7 @@ public class AidEffectivenessIndicatorUtil {
             option = (AmpAidEffectivenessIndicatorOption)
                     session.load(AmpAidEffectivenessIndicatorOption.class, optionId);
         } catch (org.hibernate.ObjectNotFoundException nfe) {
-            /*
-             * Throwing Runtime here to prevent handling Hibernate exceptions in the action class
-             * I do not want to create business level exception for this
-             */
-            throw new RuntimeException("Option with id " + optionId + " not found");
+            return null;
         }
         AmpAidEffectivenessIndicator indicator = option.getIndicator();
         indicator.getOptions().remove(option);
@@ -119,7 +142,7 @@ public class AidEffectivenessIndicatorUtil {
 
     /**
      * Returns all active (visible) indicators for the activity form to display
-     * @return
+     * @return indicators with Active = true attribute
      */
     public static List<AmpAidEffectivenessIndicator> getAllActiveIndicators() {
         Session session = PersistenceManager.getSession();
@@ -134,7 +157,7 @@ public class AidEffectivenessIndicatorUtil {
     /**
      * When indicator is physically deleted we need to clean up the reference in the amp_modules_visibility table
      * To clean up trash in the Global FM tree
-     * @param indicator
+     * @param indicator the indicator to clean up for
      */
     public static void cleanUpModulesVisibility(AmpAidEffectivenessIndicator indicator) {
         Session session = PersistenceManager.getSession();
